refactor BookForm component
