+++ src/containers/BookForm.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-unused-vars */
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
@@ -10,17 +11,9 @@ const BookForm = ({ addBook }) => {
 
   const [title, setTitle] = useState('');
   const [category, setCategory] = useState('');
-  const [id, setId] = useState(5);
-
-  const handleSubmit = e => {
-    e.preventDefault();
-    setId(id + 1);
-    addBook({ id, title, category });
-  };
 
   return (
     <form
-      onSubmit={handleSubmit}
       className="d-flex align-items-end justify-content-between shadow-top w-100 py-3 px-2 bg-light"
     >
       <div className="form-group mb-0">
