+++ src/index.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 import React from 'react';
 import ReactDOM from 'react-dom';
 import { createStore } from 'redux';
@@ -6,7 +5,7 @@ import { Provider } from 'react-redux';
 
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
-import App from './App';
+import App from './components/App';
 import rootReducer from './reducers';
 
 const initialState = {
@@ -18,11 +17,13 @@ const initialState = {
   ],
 };
 
+/* eslint-disable */
 const store = createStore(
   rootReducer,
   initialState,
   window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__(),
 );
+/* eslint-enable */
 
 ReactDOM.render(
   <React.StrictMode>
